third_party/WebKit/LayoutTests/fast/canvas/canvas-createImageBitmap-animated-expected.txt

PASS d[0] == 0 is true
PASS d[1] == 255 is true
PASS d[2] == 0 is true
PASS d[3] == 255 is true
PASS successfullyParsed is true

TEST COMPLETE
